+++ CHANGES.rst
@@ -4,8 +4,9 @@ Products.MeetingLiege Changelog
 4.2.12 (unreleased)
 -------------------
 
-- Nothing changed yet.
-
+- When finances advice is negative, automatically send the item back to the
+  internal reviewer instead sending it back to the director.
+  [gbastien]
 
 4.2.11 (2024-04-23)
 -------------------